Finish typing game course project
- Practice custom hooks by moving game state to a hook
